pyproject.toml
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = [
  "setuptools",
  "setuptools-protobuf[mypy]",
]
build-backend = "setuptools.build_meta"

[project]
name = "rock-paper-sand"
version = "0"
requires-python = ">=3.11"
dependencies = [
  "absl-py",
  "immutabledict",
  "jmespath",
  "jsonschema",
  "protobuf",
  "python-dateutil",
  "pyyaml",
  "requests",
  "requests-cache",
  "urllib3",
]

[project.optional-dependencies]
test = [
  "types-jmespath",
  "types-jsonschema",
  "types-protobuf",
  "types-python-dateutil",
  "types-pyyaml",
  "types-requests",
]

[project.scripts]
rock-paper-sand = "rock_paper_sand.main:main"

[tool.black]
line-length = 80

[tool.coverage.run]
omit = [
  "**/*_pb2.py",
  "**/*_test.py",
  "**/conftest.py",
]

[tool.isort]
profile = "google"
# TODO(https://github.com/PyCQA/isort/pull/2149): Delete this.
single_line_exclusions = [
  "collections.abc",
  "typing",
]
skip_gitignore = true
src_paths = ["rock_paper_sand"]

[tool.mypy]
disallow_any_generics = true
disallow_untyped_defs = true
packages = ["rock_paper_sand"]
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "absl.*",
  "google.protobuf.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = """\
  --cov=. \
  --cov-branch \
  --cov-report=term-missing \
  """

[tool.setuptools.packages.find]
include = ["rock_paper_sand", "rock_paper_sand.*"]
